requirements.txt
# Core
pandas>=1.5.0
numpy>=1.23.0
PyYAML>=6.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Machine Learning
scikit-learn>=1.2.0

# Association Rules Mining
mlxtend>=0.21.0

# Time Series
statsmodels>=0.13.0

# Gradient Boosting
xgboost>=1.7.0

# Data format
pyarrow>=10.0.0

# Data acquisition
kagglehub

# Demo App
streamlit>=1.30.0
plotly>=5.18.0
joblib>=1.3.0